+++ frontend/src/app/presentation/components/card-book-horizontal/card-book-horizontal.component.ts
@@ -1,9 +1,10 @@
 import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
 import { CommonModule, NgOptimizedImage } from '@angular/common';
 import { Book, BookDetail } from '../../../data/models/book';
-import { RouterLink } from '@angular/router';
+import { Router, RouterLink } from '@angular/router';
 import { DefaultButtonComponent } from '../default-button/default-button.component';
 import { CartService } from 'app/data/services/cart/cart.service';
+import { LoggedInService } from 'app/data/services/login/loggedIn.service';
 
 @Component({
   selector: 'app-card-book-horizontal',
@@ -20,15 +21,21 @@ export class CardBookHorizontalComponent {
   @Input() public className: string = '';
   @Input() public onCart: boolean = false;
 
-  constructor(public cartService: CartService) {
+  constructor(public cartService: CartService,
+    private router: Router,
+    private loggedInService: LoggedInService) {
     this.book = {} as BookDetail;
   }
 
   addOrRemove() {
-    this.onCart
+    if (this.loggedInService.isLoggedIn()) {
+      this.onCart
       ? this.cartService.deleteBookToCart(this.book)
       : this.cartService.addBookToCart(this.book);
     this.onCart = !this.onCart;
+    } else {
+      this.router.navigate(['/login']);
+    }
   }
 
   removeFromCart() {
